% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndonesiaTypoCorR.R
\name{correct}
\alias{correct}
\title{Spelling Corrector for Bahasa Indonesia}
\usage{
correct(kalimat)
}
\arguments{
\item{kalimat}{}
}
\description{
Provides function to correct spelling of a word in
Bahasa Indonesia using Naive Bayes algorithm.
}
\examples{
correc("tidakk")

}
\author{
Aisyah Zakiah
}
